refactor(driver/docker): split functions across file to reduce file size
